reduce memory usage for large aggregate states
For small requests, we simply double the array size, because that's
that AllocSet will give use anyway. No point in trying to save
memory by growing the array slower.

After reaching ALLOCSET_SEPARATE_THRESHOLD, the memory is allocated
in separate blocks, thus we can be smarter and grow the memory
a bit slower (just enough to get the 20% free space).

The performance impact is negligible - for small aggregate states
the repalloc() frequency is exactly the same as before. For large
states it's more frequent than before, but less frequent than for
small states.