Fix confusion in using PG_RETURN_DATUM in array functions
Some of the functions related to handling arrays were a bit confused
whether to use PG_RETURN_DATUM to return the result. build_array is not
a SQL function so it should be using plain return, but it was using
PG_RETURN_DATUM. On the other hand, array_agg_distinct_by_... functions
should have been using PG_RETURN_DATUM, but were not.

Fortunately, PG_RETURN_DATUM(x) just translates to "return (x)" so this
is harmless. But it's a bit confusing, because it suggests the function
is SQL callable (and vice versa).